refactor(IntentUtil): 잘못 표기되어 있는 패키지 경로를 수정
